refactor: Move config into internal
